1415 DjangoApplication1 at proj level should be called DjangoProject1  Rename the project template
 1349 extract method on global assignment doesn't preserve assignment to global if read is in inner scope
 Fixes handling of assignment in an inner scope so that we'll pass the result back out
 Also fixes handling of definite assignment of for loop variable

 1316 Breakpoint in Django view breaks on the wrong line
 We need to open the template file in binary mode so that we have correct offsets for lines

 1420 Run in Azure emulator not working
 We need to use the full path for the virtual env interpreter in our targets file

 1417 Files not in project are added to source control
 When VS queries for files that are under source control we need to also treat non-member items as not being included

 1407 Add Python WebPI reference to Django template
 Adds extra reference to template

 1277 Find in project doesn't limit itself to files in project
 Respond to IsNonSearchable from hierarchy node, treat it the same as hidden item


 1314 Sync DB should reset environment when it's done
 Reset the REPL evaluator after a command has been exectured


 1215 Can't debug django project in virtual env
 Delivering our events to the debugger is taking a long time because they are calling ReleaseComObject on our events, which aren't COM objects. This throws 2 exceptions and we're spending all of our time processing hundreds of exceptions which are occurring due to module load and other events.

 Now we marshal our events out on our own and treat them as native COM objects. This causes ReleaseComObject to succeed, and the performance problem is gone.

 This change may be controversial, so we could choose not to take this filed as DevDiv bug #724094

 1313 Django project should have more file templates in Add New Item
 Add new templates

 1312 Can't delete solution folder after closing solution
 1167 Analysis is not unloaded when a solution is closed
 We need to properly dispose of our interpreter factories. Make existing factory methods return a strongly typed disposable object, and add appropriate dispose logic for consumers

 1400 Project for analysis should not include 'Show all files' files
 When re-analyzing the project don't include non-member items

 1393 Django shell doesn't set DJANGO_SETTINGS_MODULE
 Set DJANGO_SETTINGS_MODULE when starting REPL window, and generally adds ability to set env vars


 1405 Clicking stop doesn't stop profiling
 When the wait on abnormal exit or wait on normal exit feature is turned on, we launch in a cmd wrapper That's incompatible with killing the resulting process.
 Move the waiting logic into the profiler script so that we only have a single process we need to worry about

 Also implements path expansion to help the Antares team w/ deployments to other drives
 Handle Unauthorized exceptions when updating show all files state

+++ Python/Product/VSInterpreters/ConfigurablePythonInterpreterFactoryProvider.cs
@@ -22,7 +22,7 @@ using Microsoft.VisualStudio.Shell;
 namespace Microsoft.PythonTools.Interpreter {
     [Export(typeof(IPythonInterpreterFactoryProvider))]
     class ConfigurablePythonInterpreterFactoryProvider : IPythonInterpreterFactoryProvider {
-        private readonly Dictionary<Guid, IPythonInterpreterFactory> _interpreters = new Dictionary<Guid, IPythonInterpreterFactory>();
+        private readonly Dictionary<Guid, PythonInterpreterFactoryWithDatabase> _interpreters = new Dictionary<Guid, PythonInterpreterFactoryWithDatabase>();
         private readonly SettingsManager _settings;
 
         // keys used for storing information about user defined interpreters
@@ -41,7 +41,7 @@ namespace Microsoft.PythonTools.Interpreter {
             DiscoverInterpreterFactories();
         }
 
-        private IPythonInterpreterFactory LoadUserDefinedInterpreter(SettingsStore store, string guid) {
+        private PythonInterpreterFactoryWithDatabase LoadUserDefinedInterpreter(SettingsStore store, string guid) {
             // PythonInterpreters\
             //      Id\
             //          Description
@@ -80,16 +80,25 @@ namespace Microsoft.PythonTools.Interpreter {
         }
 
         public void RemoveInterpreter(Guid id) {
-            var collection = PythonInterpreterKey + "\\" + id.ToString("B");
-            var store = _settings.GetWritableSettingsStore(SettingsScope.UserSettings);
+            PythonInterpreterFactoryWithDatabase fact;
+            if (_interpreters.TryGetValue(id, out fact)) {
+                var collection = PythonInterpreterKey + "\\" + id.ToString("B");
+                var store = _settings.GetWritableSettingsStore(SettingsScope.UserSettings);
 
-            store.DeleteCollection(collection);
-            _interpreters.Remove(id);
-            OnInterpreterFactoriesChanged();
+                store.DeleteCollection(collection);
+
+                _interpreters.Remove(id);
+                OnInterpreterFactoriesChanged();
+                fact.Dispose();
+            }
         }
 
         public bool IsConfigurable(IPythonInterpreterFactory factory) {
-            return _interpreters.ContainsValue(factory);
+            PythonInterpreterFactoryWithDatabase fact = factory as PythonInterpreterFactoryWithDatabase;
+            if (fact == null) {
+                return false;
+            }
+            return _interpreters.ContainsValue(fact);
         }
 
         public IPythonInterpreterFactory SetOptions(InterpreterFactoryCreationOptions options) {
@@ -107,6 +116,10 @@ namespace Microsoft.PythonTools.Interpreter {
             var newInterp = LoadUserDefinedInterpreter(store, options.IdString);
             Debug.Assert(newInterp != null);
 
+            PythonInterpreterFactoryWithDatabase existing;
+            if (_interpreters.TryGetValue(newInterp.Id, out existing)) {
+                existing.Dispose();
+            }
             _interpreters[newInterp.Id] = newInterp;
             OnInterpreterFactoriesChanged();
             return newInterp;
@@ -139,7 +152,9 @@ namespace Microsoft.PythonTools.Interpreter {
             }
 
             foreach (var id in notFound) {
+                var existing = _interpreters[id];
                 _interpreters.Remove(id);
+                existing.Dispose();
             }
 
             if (anyChange) {
